Enemy movement
Her variables are walkPoints (list of Vector2), speed (float) and move (bool). The walkPoints is a places where the enemy will go. The boolean move it's true only if the enemy needs to be always walking, if he has other behaviours, this boolean need to be false, and to enemy walk, other code need to call Move() method. I added a example of enemy in SampleScene. Any doubt feel free to ask me