Enemy follow
I restructured the enemy codes, and now the enemy has a movementCotroller, patrol and follow scripts. The first its a global code for movements, has method to follow object, speed variable, etc. The second its the code who do enemy move, and the third do enemy follow the player when he its on the vision radius, it has a three variables: visionRadius (Vector2), minTimeToStop and maxTimeToStop (floats). The timeToStop variables are used to enemy has a delay to stop follow when player exit of her visionRadius (idea of Maki), that delay its a random number between min and max time to stop